backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-genai
pydantic>=2.0
slowapi
python-multipart
aiofiles
ffmpeg-python
torch
faster-whisper

# ── NEW: AI Insights Dashboard ─────────────────────────────────────────────────
# Core ML (lightweight — no heavy torch dependency)
scikit-learn>=1.4.0          # StandardScaler, train_test_split
xgboost>=2.0.3               # No-show prediction model (production path)
prophet>=1.1.5               # Revenue time-series forecast (production path)
 
# Data wrangling
pandas>=2.2.0
numpy<2
 
# Model persistence
joblib>=1.3.2
 